requirements.txt
RPi.GPIO==0.7.0
pydub==0.25.1
requests==2.25.1
python-dotenv==0.21.0
